Experimental view distance modification
This is just a rough idea of how we might implement a way of increasing
the view distance. It has some hackish code and still a few bugs but it
can give you an idea of what can be done with it. Pushing just so you
guys can see and maybe play around with. Very much WIP.